Feature/improve test coverage

+++ test/test_airtunnel/conftest.py
@@ -38,7 +38,8 @@ def test_db_path() -> str:
     logger.info(f"Using test-database: {db_path}")
     yield db_path
     logger.info(f"Removing file of test-database: {db_path}")
-    os.remove(db_path)
+    if os.path.exists(db_path):
+        os.remove(db_path)
 
 
 @pytest.fixture(scope="session", autouse=True)
@@ -52,9 +53,7 @@ def provide_airflow_cfg(test_db_path: str) -> None:
         test_folder_path, "airflow_home/airflow.template.cfg"
     )
 
-    dags_folder_path = os.path.join(
-        test_folder_path, "test_airtunnel/workflows/testdags"
-    )
+    dags_folder_path = os.path.join(test_folder_path, "test_airtunnel/testdags")
 
     decls_folder_path = os.path.join(test_folder_path, "test_airtunnel/declarations")
